requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
mysqlclient==2.2.0
Pillow==10.1.0
python-decouple==3.8
gunicorn==21.2.0
whitenoise==6.6.0
django-storages==1.14.2
boto3==1.34.0
celery==5.3.4
redis==5.0.1
django-crispy-forms==2.1
crispy-bootstrap5==0.7
django-extensions==3.2.3
django-debug-toolbar==4.2.0
django-cleanup==8.0.0
django-ckeditor==6.7.0
django-allauth==0.57.0
django-ratelimit==4.1.0
django-simple-captcha==0.6.0
django-sitemaps==1.0.0
django-robots==5.0
django-meta==3.0.0
django-seo-js==2.0.0
django-compressor==4.4
django-livereload-server==0.3.2
django-environ==0.11.2
psycopg2-binary==2.9.9
dj-database-url==2.1.0
django-heroku==0.3.1
